Add files via upload
Stages of the To-Do List project. 

Part 1:  A button that adds text to the page. 
Part 2:  An input field, and a button that adds that text to the page (repeatedly).  Each new item added to the list has an onclick function that allows you to remove that item.  
Part 3:  Added CSS style to the page.  The JavaScript is where we left off with Part 2.  Two new buttons are added to the HTML (Save and Load) but they need their JavaScript code added.